refactor: drop redundant int() around round()
`round(x)` with no ndigits already returns an int; only `round(x, n)` returns
a float. The wrappers in framebuffer.py and infrared.py were no-ops.

These were the last two occurrences in the tree; the ones in color.py and
aurora.py were removed alongside the hue quantisation fix.

+++ src/lifx/devices/infrared.py
@@ -176,7 +176,7 @@ class InfraredLight(Light):
             )
 
         # Convert from float (0.0-1.0) to uint16 (0-65535)
-        brightness_u16 = max(0, min(65535, int(round(brightness * 65535))))
+        brightness_u16 = max(0, min(65535, round(brightness * 65535)))
 
         # Request automatically handles acknowledgement
         result = await self.connection.request(
